split loop code from state.rs into loop_step.rs
